implement individual trace encoding
