+++ ddtrace/encoding.py
@@ -144,36 +144,4 @@ class MsgpackEncoder(_EncoderBase):
             return struct.pack('>BI', 0xdd, count) + buf
 
 
-
-from msgpack.fallback import Packer
-
-class PPMsgpackEncoder(_EncoderBase):
-    content_type = 'application/msgpack'
-
-    @staticmethod
-    def encode(obj):
-        return Packer().pack(obj)
-
-    @staticmethod
-    def decode(data):
-        if msgpack.version[:2] < (0, 6):
-            return msgpack.unpackb(data)
-        return msgpack.unpackb(data, raw=True)
-
-    @staticmethod
-    def join_encoded(objs):
-        """Join a list of encoded objects together as a msgpack array"""
-        buf = b''.join(objs)
-
-        # Prepend array header to buffer
-        # https://github.com/msgpack/msgpack-python/blob/f46523b1af7ff2d408da8500ea36a4f9f2abe915/msgpack/fallback.py#L948-L955
-        count = len(objs)
-        if count <= 0xf:
-            return struct.pack('B', 0x90 + count) + buf
-        elif count <= 0xffff:
-            return struct.pack('>BH', 0xdc, count) + buf
-        else:
-            return struct.pack('>BI', 0xdd, count) + buf
-
-
 Encoder = MsgpackEncoder
